#9 wrapper helpers
